+++ src/Phobos.Ext.cpp
@@ -19,6 +19,7 @@
 #include <Ext/Techno/Body.h>
 #include <Ext/TechnoType/Body.h>
 #include <Ext/TerrainType/Body.h>
+#include <Ext/Tiberium/Body.h>
 #include <Ext/VoxelAnim/Body.h>
 #include <Ext/VoxelAnimType/Body.h>
 #include <Ext/WarheadType/Body.h>
@@ -242,6 +243,7 @@ auto MassActions = MassAction <
 	TechnoExt,
 	TechnoTypeExt,
 	TerrainTypeExt,
+	TiberiumExt,
 	VoxelAnimExt,
 	VoxelAnimTypeExt,
 	WarheadTypeExt,